add css module build support for react
