+++ packages/builder-config/framework-css/index.ts
@@ -1,0 +1,11 @@
+export type CssStylingConfig = VanillaCssStylingConfig | ScssStylingConfig;
+
+interface VanillaCssStylingConfig {
+  type: "css";
+  lang: "css";
+}
+
+interface ScssStylingConfig {
+  type: "css";
+  lang: "scss";
+}
