+++ packages/builder-config/framework-react/react-config-styling.ts
@@ -1,24 +1,16 @@
+import type { CssStylingConfig } from "../framework-css";
+
 export type ReactStylingStrategy =
   | ReactInlineCssConfig
-  | ReactCssStylingConfig
+  | ReactNormalCssConfig
+  | ReactCssModuleConfig
   | ReactStyledComponentsConfig;
 
 export type ReactStyledComponentsConfig =
   | ReactTheStyledComponentsConfig
   | ReactEmotionStyledConfig;
-export type ReactCssStylingConfig = CssStylingConfig;
 
-type CssStylingConfig = VanillaCssStylingConfig | ScssStylingConfig;
-
-interface VanillaCssStylingConfig {
-  type: "css";
-  lang: "css";
-}
-
-interface ScssStylingConfig {
-  type: "css";
-  lang: "scss";
-}
+export type ReactNormalCssConfig = CssStylingConfig;
 
 /**
  * "The" styled-components config - https://styled-components.com/
@@ -48,3 +40,42 @@ interface ReactEmotionStyledConfig {
 export interface ReactInlineCssConfig {
   type: "inline-css";
 }
+
+/**
+ * Configuration for css module for react
+ *
+ *
+ * What is css module?
+ * ```tsx
+ * import styles from "./component.module.css";
+ * <div className={styles.box}/>
+ * ```
+ */
+export type ReactCssModuleConfig = Omit<CssStylingConfig, "type"> & {
+  type: "css-module";
+
+  /**
+   * The name of the css module
+   * @deprecated not used
+   *
+   * - .css
+   * - .module.css
+   */
+  pattern?: string;
+
+  /**
+   * the identifier of the css module to use when importing
+   *
+   * ```
+   * import <importDefault> from "./component.module.css";
+   * ```
+   */
+  importDefault: string | "styles";
+
+  /**
+   * Loader used for webpack to enable css module feature
+   *
+   * https://github.com/webpack-contrib/css-loader
+   */
+  loader: "css-loader";
+};
